+++ src/main/kotlin/models/Tag.kt
@@ -3,4 +3,4 @@ package org.example.models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Tag(val id: Int? = null, val name: String? = null)
+data class Tag(val id: Long? = null, val name: String? = null)
